Finishing Edits
Cleaned up code based on Jacks comments
