+++ app/config/routes.js
@@ -5,7 +5,6 @@ var Route = ReactRouter.Route;
 var IndexRoute = ReactRouter.IndexRoute;
 var hashHistory = ReactRouter.hashHistory;
 var Main = require('../components/Main');
-var Home = require('../components/Home');
 var BarPage = require('../containers/BarPage');
 var Menu = require('../containers/Menu');
 var Homebrew = require('../containers/Homebrew');
@@ -13,12 +12,10 @@ var Homebrew = require('../containers/Homebrew');
 var routes = (
 
   <Router history={hashHistory}>
-    <Route path='/' component={Main}>
-      <IndexRoute component={Home} />
+      <Route path='/' component={Main}/>
       <Route path='/homebrew' component={Homebrew}/>
       <Route path='/barpage' component={BarPage}/>
       <Route path='/menu' component={Menu}/>
-    </Route>
   </Router>
 );
